fix codegen tests: migrate PGType/BaseType from string to typeinfo.Type
Update all test files in internal/codegen/ to use typeinfo.MustParse()
for PGType and BaseType struct fields, which now require typeinfo.Type
instead of string. MustParse normalizes type aliases (e.g. bigint->int8,
integer->int4, boolean->bool) to match production code paths.

Domain-backed columns use typeinfo.Type{Base, DomainName} directly.
Golden file ddl_expected.py updated for normalized type names.
GORM tag expectations updated for canonical short-form types.

Claude-Code-Session-Id: df7f98a7-e6d9-4c23-aae2-b5db6ec5707d

+++ internal/codegen/testdata/ddl_expected.py
@@ -17,7 +17,7 @@ STATEMENTS: Final[list[tuple[str, str, str | None, int]]] = [
     ("""CREATE TABLE shop.orders (
     id uuid NOT NULL DEFAULT gen_random_uuid(),
     customer_id uuid NOT NULL,
-    total bigint NOT NULL DEFAULT 0,
+    total int8 NOT NULL DEFAULT 0,
     created_at timestamptz NOT NULL DEFAULT now(),
     CONSTRAINT pk_orders PRIMARY KEY (id)
 );""", "table", "orders", 4),
